requirements.txt
SQLAlchemy
alembic
PyQt5
cryptography
PyJWT
pytest